iotune: ignore shards with id above max_iodepth
If max_iodepth is less than smp::count, then the
calculated io_depth_per_shard can be zero for shards
with id that is greater or equal to the max_iodepth.
This causes runtime error during random read and
random write measurements.

Such sitation was encountered when iotune was run on
a machine with 128 CPUs and non-RAID disk that showed
max_iodepth=127.

This change introudces a logic that prints a warning in
such case and skips performing measurements on shards
that would have io_depth_per_shard=0.

Refs: scylladb#2169

Signed-off-by: Patryk Wrobel <patryk.wrobel@scylladb.com>

Closes scylladb/seastar#2201