fair_queue: Scale fixed-point factor
The vaulue is used to convert request tokens from float-point number to
some "sane" integer. Currently on ~200k IOPS ~2GBs disk it produces the
following token values for requests of different sizes

    512:    150k
    1024:   160k
    2048:   170k
    ...
    131072: 1.4M

These values are pretty huge an when accumulated even in 64-bit counter
can overflow it in months time scale. Current code sort of accounts for
that by checking the overflow and resetting the counters, but in the
future there will be the need to reset counters on different shards, and
that's going to be problematic.

This patch reduces the factor 8 times, so that the costs are now

    512:      19k
    1024:     20k
    2048:     21k
    ...
    131072:  170k

That's much more friendly to accumulating counters (the overflow is now
at the year's scale which is pretty comfortable). Reducing it even
further is problematic, here's why.

In order to provide cross-class fairness the costs are divided by class
shares for accumulation. Given a class of 1000 shares, the 512-bytes
request becomes indistinguishable from 1k one with smaller factor. Said
that, even with the new factor it's worth taking more care when dividing
the cost at shares use div-roundup math.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>