io_queue: Stall detector
Detect requests with execution delay larger than the configured
threshold, print warning into logs that includes the number of queued
and executing requests for the queue.

Other than that, include the number of polls that happened while the
request was in disk. For that, the executing request notices the
reactor::_polls count when it gets dispatched. This metrics should help
telling disk stalls from reactor misbehaviors. If the request wasn't
complete for too long due to disk/kernel problems, the polls count
difference should corresond to the observed delay: polls times latency
goal should be comparable to or greater than the request delay. Too low
polls count would mean that reactor failled to fetch the completion from
the kernel.

The threshold is doubled every time stall is detected to avoid storm of
warning. The averaging timer (introduced in previous patch) slowly
decreases the threshold back to keep catching "short" stalls. Initial
threshold is set to be 10x the io latency goal.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>