Delayed accumulator fixup on IO class activation
When a class receive a request into empty queue it's "activated".
Activation means that the class is queued for scheduling and at that
time it may have its accumulator counter being very low, thus getting
unwanted priority against other classes. To prevent that, the
accumulator is bumped up slightly below the top-most class in the queue.

This bump-up looks to be pretty aggressive, interactive classes like
statement one constantly lose big portion of their accumulator thus
losing the ability to consume disk capacity at the rate they want.

This patch delays the bump-up until the class is scheduled and
dispatched from, thus giving it ability to drain as much accumulated
preference as it could.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>